+++ electron-app/src/icons/firefly-icon-ascii.txt
@@ -1,0 +1,52 @@
+......................................................................................................
+........................................###########.....................########......................
+....................................########..........-#######...............######...................
+.................................#######.....########################...........#####.................
+..............................######....###############.......##########...........###................
+...........................######...####..###...###################..######.........####..............
+.........................#####..########..#############################..####.........###.............
+......................#####..#######.+##..####..###############....#######..##..........##............
+....................#####.######..#######.#########............####.....####.##.........##............
+..................####..#####..#####..######...################......#.....##.##.........##...........
+.................###..####...####.######.##########............####...+.....#####-........##..........
+...............###..####..####..####.#######........................##........####.........##.........
+.............####.####..###..###..######.......................................#.#-........##.........
+............###.###+..###.####..####............................................#.#........##.........
+..........-##.####..###.###..####...............................................#.#........##.........
+.........##..###...##..##.#####............................................................##.........
+........##.###...##....#####...............................................+++.............##.........
+.......##.##....##.##-..##.............-###########################........................##.........
+......##.##...+#..#...................###.............-------++##.##.......++..............##.........
+.....#..##...#...#.####.............###............................#.......................##.........
+....#..##...#..+.####..............##..............................#.......................##.........
+...#.##.......###.................##........########################......................##..........
+..#..#......#+..##................##.......##.............................................##..........
+.#..#..+###....##.................##.......##.............................................##..........
+.#.#...#-.#...##..................##........###################..........................##.........#.
+..#....####..##...................##........-###################........................##..........#.
+.#......##..##....................##..........................-#.......................##.....##...##.
+.#..........#.....................##..........................-#......................##..........#+#.
+...........##.....................##........-###################.....................##.......##.##+#.
+..........##......................##........###################.....................##..........##.#..
+..........##......................##.......##.....................................###......###.##.#...
+.........##.......................##.......##....................................###..........##.##...
+.........##.......................##.......##..................................###.........#...###....
+.........##.......................##.......##.................................###..........##.##......
+.........##.......................+#.--..-.#-...............................###...##.....###.##.......
+.........##....+++.................-#######-..............................####...#-..##.###.##........
+.........##.............................................................####...##..+...##.###.........
+.........##....+++...................................................#####...##..#+....#.###..........
+.........##........-.-............................................#####...###..##..###..##............
+.........##..................##...............................#######..####..##..###..###.............
+..........##..........#........###........................#######+..####..###.####..###...............
+..........##.........#.##....#+...#####............##.#######-...####..####.####..###.................
+...........##.........#.###....###...################.--##..#####+..####..####..####..................
+............##.........#..####....#####.............#.#######-...####..#####.#####....................
+.............##.........##..#######..+###############.......#####..######..#####......................
+..............##.........####......##########......##########..#######..#####.........................
+...............###.........-#####..+###################..#########...######...........................
+................####..........###..-####-......##############.....######..............................
+..................####............######################......#######.................................
+....................#####.................+##.............#######+....................................
+......................########.....................###########........................................
+......................................................................................................
