+++ electron-app/src/main-utils/state.js
@@ -1,0 +1,13 @@
+// state variables that will be shared
+module.exports = {
+    logWindow: null,
+    notebookPath: null,
+    jupyterPath: null,
+    pythonPath: null,
+    mainWindow: null,
+    splash: null,
+    pyProc: null,
+    fireflyPort: null,
+    jupyterProc: null,
+    jupyterPort: null,
+};
