+++ electron-app/src/webviews/firefly-electron.html
@@ -10,16 +10,16 @@
         <div id="container" style="display: flex; height: 100vh; width: 100vw">
             <div id="sidebar">
                 <div class="sidebar-icon-wrapper">
-                    <button onclick="toggleView(this, 'firefly')" class="shown"><img width="60px" src="icons/firefly-icon.png"></button>
+                    <button onclick="toggleView(this, 'firefly')" class="shown"><img width="60px" src="../icons/firefly-icon.png"></button>
                 </div>
                 <div class="sidebar-icon-wrapper">
-                    <button onclick="toggleView(this, 'jupyter')" class="shown"><img width="60px" src="icons/jupyter-icon.png"></button>
+                    <button onclick="toggleView(this, 'jupyter')" class="shown"><img width="60px" src="../icons/jupyter-icon.png"></button>
                 </div>
                 <div class="sidebar-icon-wrapper" id="loadDataButton">
-                    <button onclick="enableDataTypeSelection()"><img width="60px" src="icons/Google_Design_folder_upload_icon.png"></button>
+                    <button onclick="enableDataTypeSelection()"><img width="60px" src="../icons/Google_Design_folder_upload_icon.png"></button>
                 </div>
                 <div class="sidebar-icon-wrapper">
-                    <button onclick="toggleView(this, 'logs')"><img width="60px" src="icons/Google_Design_overview_icon.png"></button>
+                    <button onclick="toggleView(this, 'logs')"><img width="60px" src="../icons/Google_Design_overview_icon.png"></button>
                 </div>
 
             </div>
