+++ electron-app/src/webviews/firefly-electron.js
