+++ electron-app/src/webviews/splash.html
@@ -28,7 +28,7 @@
 </head>
 <body>
     <div class="content">
-        <img src="icons/firefly-banner.png" width="720px" alt="Firefly banner">
+        <img src="../icons/firefly-banner.png" width="720px" alt="Firefly banner">
         <h1>Loading...</h1>
     </div>
 </body>
